Modify pool stat to take an overlay component, make hovering over any part of the Pool Row stat trigger the overlay
