+++ apps/hyperdrive-trading/src/ui/markets/PoolRow/PoolStat.tsx
@@ -1,11 +1,11 @@
+import * as Tooltip from "@radix-ui/react-tooltip";
 import classNames from "classnames";
 import { ReactElement, ReactNode } from "react";
 import Skeleton from "react-loading-skeleton";
-import { Tooltip } from "src/ui/base/components/Tooltip/Tooltip";
 
 export interface PoolStatProps {
   label: string;
-  labelTooltip?: string;
+  overlay?: ReactNode;
   value: ReactNode;
   isLoading?: boolean;
   isNew?: boolean;
@@ -15,8 +15,8 @@ export interface PoolStatProps {
 
 export function PoolStat({
   label,
-  labelTooltip,
   value,
+  overlay,
   isNew = false,
   variant = "default",
   isLoading = false,
@@ -31,17 +31,9 @@ export function PoolStat({
     displayValue = value;
   }
 
-  return (
+  const poolStat = (
     <div className="flex shrink-0 flex-col items-center gap-1.5 sm:items-start">
-      <Tooltip
-        className={
-          "group text-neutral-content before:z-40 before:max-w-56 before:text-start sm:text-sm"
-        }
-        tooltip={labelTooltip}
-      >
-        {label}
-      </Tooltip>
-
+      <div className={"text-neutral-content"}>{label}</div>
       <div
         className={classNames("font-dmMono text-h4 font-medium", {
           "gradient-text": variant === "gradient",
@@ -52,4 +44,27 @@ export function PoolStat({
       {action}
     </div>
   );
+
+  if (overlay) {
+    return (
+      <Tooltip.Provider>
+        <Tooltip.Root>
+          <Tooltip.Trigger className="flex items-center whitespace-nowrap">
+            {poolStat}
+          </Tooltip.Trigger>
+          <Tooltip.Portal>
+            <Tooltip.Content
+              className="z-20 h-fit w-72 rounded-[0.6rem] bg-base-200 px-3 py-2 shadow-2xl"
+              sideOffset={5}
+              collisionPadding={12}
+            >
+              {overlay}
+            </Tooltip.Content>
+          </Tooltip.Portal>
+        </Tooltip.Root>
+      </Tooltip.Provider>
+    );
+  }
+
+  return poolStat;
 }
